Moving test files out of pages directory
