+++ __test__/pages/Home.test.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { render, screen } from "@testing-library/react"
 
-import Home from "./index"
+import Home from "../../pages/index"
 
 describe("Home", () => {
   it("should render an H2", () => {
